update rov behavior, before 2019-8-13 exp

+++ include/detector.h
@@ -32,8 +32,7 @@ private:
     bool track;
     unsigned char track_cl;
     int track_id;
-    unsigned long long frame_num;
-    //    unsigned feature_size;
+    unsigned int frame_num;
     std::vector<cv::Scalar> color{cv::Scalar(0,0,0), cv::Scalar(255,255,0), cv::Scalar(0, 255,255), cv::Scalar(0,0,255), cv::Scalar(255,0,255)};
 public:
     Detector(unsigned int, int, float, unsigned char, int, bool);
